Module_7/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1

# Celery + Redis (task queue)
celery==5.3.6
redis==5.0.4

# PDF generation (xhtml2pdf or WeasyPrint, you can use either)
xhtml2pdf==0.2.15
jinja2==3.1.3

